Move close-during-CONNECTING tests to their own file
websocket.test.js has several pre-existing tests (300× TLS handshakes,
1000× connect+close, external echo host) that run near the 5s default
timeout under debug+ASAN regardless of this change. Keeping the new
tests in a dedicated file — consistent with the other per-topic test
files in this directory — lets them be evaluated without those
unrelated borderline tests.
